fix errors on dashboard page and sign in page
When not signed in, the dashboard button would navigate to the
sign in screen, but the browser's back button would not go back to the
home screen. The sign in button on the sign in page was also fixed so
that it continues to display the loading animation until the user is
redirected for authentication.